% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptp.R
\name{parallel.trends.data}
\alias{parallel.trends.data}
\title{Creates a data frame used for generating plots to diagnose parallel trends for difference-in-difference regressions with additional control variables}
\usage{
parallel.trends.data(
  df,
  timevar = "t",
  yvar = "y",
  treatdummy = "treat",
  expdummy = "exp",
  treat_exp_dummy = NULL,
  cvars = NULL,
  extravars = NULL,
  add.no.control = TRUE,
  constant.type = c("means", "zero")[1]
)
}
\arguments{
\item{df}{The data frame used for the DID regression}

\item{timevar}{Name of the time variable, e.g. "year"}

\item{yvar}{Name of the dependent variable}

\item{treatdummy}{Name of the 0-1 dummy variable indicating whether the observation belongs to the treatment group}

\item{expdummy}{Name of the 0-1 dummy variable indicating whether the observation is in an experimental period, i.e. a period in which the treatment group gets treatment.}

\item{treat_exp_dummy}{Optional name of the interaction variable treatdummy * expdummy. If not given a column with name "treat_exp_dummy" is added to the data frame.}

\item{cvars}{Names of the additional control variables used in the DID regression.}

\item{extravars}{Names of variables that are not used in the regression but shall be included in the returned data frame.}

\item{add.no.control}{If TRUE (default) add rows that are used for a parallel trends diagnostic plot that assumes that no control variables are added.}
}
\value{
A data frame that can be passed to parallel.trends.plot. Essential columns are .t, .y, .group, .exp, .mode. Note that the data frame only contains rows that have no NA for the specified columns.
}
\description{
The resulting data frame can be directly passed to parallel.trends.plot or be used to generate a more customized manual plot.
}
