% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptp.R
\name{parallel.trends.plot}
\alias{parallel.trends.plot}
\title{Generates plots to diagnose parallel trends for difference-in-difference regressions with additional control variables}
\usage{
parallel.trends.plot(
  pt.dat,
  facet.mode = ".mode" \%in\% colnames(pt.dat),
  add.exp.line = c("first", "all", "none")[2],
  exp.line.opt = list(color = "black")
)
}
\arguments{
\item{pt.dat}{A data frame as returned by the function parallel.trends.data}

\item{facet.mode}{If TRUE generate a facet plot with one facet for trends of control group and experimental group with and without correction of the control variables}

\item{add.exp.line}{How shall vertical lines be added to indicate the experimental periods? "first" means a line at the first experimental period, "all" means a line at all experimental periods. Use "none" to add no lines}

\item{exp.line.opt}{Arguments passed to geom_vert to customize the vertical lines.}
}
\value{
A ggplot object
}
\description{
Generates plots to diagnose parallel trends for difference-in-difference regressions with additional control variables
}
